requirements.txt
pypresence
pystray
psutil
Pillow
cursor
InquirerPy